perf(nodes-routing): delete, settings, and edit dialog utilize the same central cache
